Fix compiler warnings
- uint64_t for ui64 typedef
- Format specifier usage for ui64 printing

+++ bitboard.h
@@ -1,8 +1,9 @@
 #ifndef BITBOARD_H
 #define BITBOARD_H
 
+#include <inttypes.h>
 
-typedef unsigned long long ui64;
+typedef uint64_t ui64; // unsigned long long
 
 #define get_bit(bitboard, square) ((bitboard) & (1ULL << (square)))
 #define set_bit(bitboard, square) ((bitboard) |= (1ULL << (square)))
